Fixed FAQ formatting and added title and date late updated.
Both EWW and Lynx render the title at the top of the screen.
But it seemed slightly less obvious than I had hoped, so I added an explicit title.
And I figured it would be nice to have a "last updated" tag below the title.
I may move the last updated date to the bottom of the page.
And I messed up on some of the formatting on the FAQ, so I fixed that.